Optimize upb's encoding of the packed status of packable repeated fields
Upb minidescriptors encode a default packed status for each message, and then
each packable repeated field within it can indicate if it deviates from the
default. Currently the default is determined by the `repeated_field_encoding`
feature, which means that an upgrade from proto2 to editions will always change
the minidescriptor. This makes it difficult to output minidescriptors in Rust
generated code, because we have tests that verify that an editions upgrade does
not result in generated code changes.

This CL fixes the problem by determining the default packed status in a way
that does not depend on editions. Instead we just consider packed the default
if there are more packed fields than unpacked among packable repeated fields.
This has the side benefit of somewhat reducing the size of minidescriptors for
many messages.

PiperOrigin-RevId: 761141535